Split up OrderService to exclude the statistics
